Delete gcda and gcno files on make clean.
Cmake does not know about the gcda and gcno files generated by the compiler.
With this patch they will be added to the clean target, so they will be cleaned
on make clean. Beware that this patch can't clean all files, due we can't know,
if a file includes e.g. a header for which a second gcov file will be generated.